+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 The changes documented here do not include those from the original repository.
 
+## [Version 2.3.0]
+
+### Features
+- Enable message delivery metrics exportation to BigQuery (https://outsystemsrd.atlassian.net/browse/RMET-3511).
+
+### Fixes
+- Remove the main thread blocking (https://outsystemsrd.atlassian.net/browse/RMET-3508).
+- Update `oscordova` dependency (https://outsystemsrd.atlassian.net/browse/RMET-3540)
+
 ## [Version 2.2.1]
 
 ### 03-06-2024